moved prometheus query and config controller to their own files and updated imports and exports
